refactored base notes class
